Replace anyhow with structured errors (#2)
Use a structured errors instead of just a string when accessing IPLD Lists/Maps.
Also don't error if the key does not exist, but return an option intead. This
is better aligned with `get` functions in Rust's standard library.

Fix two Serde tests as the behaviour of the deserializer was changed, but the
tests weren't updated.

Due to no needing anyhow anymore makes the conversion work with `no_std`.