chore: use local version when running tests
This patching is a hack in order to make the rustdoc tests and releases happy.
We include README in the library docs, this way the are run as tests. Those
examples create a circular dependency on `ipld-core` (as `serde_ipld_dagcbor`
and `serde_ipld_dagjson` depend on `ipld-core`.
Also without this change `cargo release` would complain, as the `Cargo.lock`
needs modifications as the previously mentioned crates would need to depend on
an already released and not the about to be released version of `ipld-core`.