chore: update dev dependencies (#15)
Update the dev dependencies to the newest version so that the tests
pass again. This is an issue whenever there is a breaking release
of `ipld-core`. On the other hand it makes sure that we won't forget
to update `serde_ipld_dagcbor` and `serde_ipld_dagjson` to the latest
version of `ipld-core` after a release.

It also adds a CI job to make sure the checked-in Cargo.lock file is
up-to-date.