feat: implement Eq for Ipld (#17)
Implement the `Eq` trait for the `Ipld` enum. NaN floats are never equal
to another. Though in the Ipld Data model they are forbidden. This means
that we can implement `Eq` for it.

As it's currently not strictly enforced that an `Ipld::Float` does not
contain a NaN, we make at least sure that in case we would encounter it,
that it is treated as equal.

For more information about the dicussion about this topic, see
https://github.com/ipld/libipld/issues/171.